Fix build on x86_64

+++ src/Application1.cpp
@@ -1853,7 +1853,7 @@ bool Application1::GetImageSize()
 				imgsize = GetBFSImageSize(base_folder.String());
 
 				window1->view1->box2->SetCDSize((off_t)imgsize);
-				delete ((BString*)list->RemoveItem((long int)0));
+				delete ((BString*)list->RemoveItem((long int)0l));
 				delete list;
 				return true;
 			}
@@ -1892,7 +1892,7 @@ bool Application1::GetImageSize()
 				imgsize = GetBFSImageSize(base_folder.String());
 
 				window1->view1->box2->SetCDSize((off_t)imgsize + sz);
-				delete ((BString*)list->RemoveItem((long int)0));
+				delete ((BString*)list->RemoveItem((long int)0l));
 				delete list;
 				return true;
 			}
@@ -1910,7 +1910,7 @@ bool Application1::GetImageSize()
 				imgsize = GetBFSImageSize(base_folder.String());
 
 				window1->view1->box2->SetCDSize((off_t)imgsize);
-				delete ((BString*)list->RemoveItem(0L));
+				delete ((BString*)list->RemoveItem((long int)0l));
 				delete list;
 				return true;
 			}
@@ -2514,7 +2514,7 @@ void Application1::DoAudioMenu()
 	window1->Lock();
 
 	while (window1->view1->listview1->cmenu->CountItems() > 0)
-		window1->view1->listview1->cmenu->RemoveItem(0L);
+		window1->view1->listview1->cmenu->RemoveItem((long int)0l);
 	window1->view1->listview1->cmenu->AddItem(window1->view1->listview1->deleteItem);
 	window1->view1->listview1->cmenu->AddSeparatorItem();
 	window1->view1->listview1->cmenu->AddItem(window1->view1->listview1->moveupItem);
@@ -2534,7 +2534,7 @@ void Application1::DoDataMenu()
 	window1->Lock();
 
 	while (window1->view1->listview1->cmenu->CountItems() > 0)
-		window1->view1->listview1->cmenu->RemoveItem(0L);
+		window1->view1->listview1->cmenu->RemoveItem((long int)0l);
 	window1->view1->listview1->cmenu->AddItem(window1->view1->listview1->deleteItem);
 	window1->view1->listview1->cmenu->AddItem(window1->view1->listview1->newfolderItem);
 	window1->view1->listview1->cmenu->AddItem(window1->view1->listview1->renamefileItem);
@@ -4514,7 +4514,7 @@ void Application1::ShowRecentProjects()
 
 	window1->Lock();
 	while (window1->view1->openrecentSM->CountItems() > 0)
-		delete window1->view1->openrecentSM->RemoveItem(0L);
+		delete window1->view1->openrecentSM->RemoveItem((long int)0l);
 	for (int i = 0; i < CM->CountStrings(RECENTPROJECTS_LIST); i++) {
 		if (CM->GetStringI(RECENTPROJECTS_LIST, i)->Length() > 0) {
 			// add if project exists, only
